requirements.txt
google-genai>=1.0.0               # Gemini API - NEW SDK (replaces google-generativeai)
pypdf>=4.0.0                      # PDF text extraction
python-docx>=1.1.0                # DOCX text extraction
python-dotenv>=1.0.0              # loads .env file safely
rich>=13.0.0                      # pretty terminal output
chromadb>=0.5.0                   # local vector database (Phase 3)
fastapi>=0.110.0                  # API framework (Phase 6)
uvicorn>=0.29.0                   # ASGI server for FastAPI (Phase 6)
streamlit>=1.35.0                 # UI framework (Phase 6)
requests>=2.31.0                  # HTTP client for UI → API calls (Phase 6)